requirements.txt
flask
python-dotenv
psycopg2-binary
bcrypt
pyjwt
datetime
flask_migrate